# api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ACCESS Ready API - Adaptive ELL exit prep for K-12 students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /districts:
    get:
      tags: [districts]
      operationId: listDistricts
      summary: List all districts
      responses:
        "200":
          description: List of districts with school counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/District"
    post:
      tags: [districts]
      operationId: createDistrict
      summary: Create a district
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDistrictBody"
      responses:
        "201":
          description: Created district
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/District"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /districts/{districtId}:
    get:
      tags: [districts]
      operationId: getDistrict
      summary: Get a district by ID
      parameters:
        - name: districtId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: District
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/District"
        "404":
          description: District not found
    patch:
      tags: [districts]
      operationId: updateDistrict
      summary: Update a district
      parameters:
        - name: districtId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDistrictBody"
      responses:
        "200":
          description: Updated district
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/District"
        "404":
          description: District not found
  /districts/{districtId}/schools:
    get:
      tags: [districts]
      operationId: listDistrictSchools
      summary: List schools in a district
      parameters:
        - name: districtId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Schools in this district
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
...
  /media:
    get:
      tags: [media]
      summary: List media assets
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of media assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAssetList"
    post:
      tags: [media]
      summary: Register a media asset after S3 upload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaAssetBody"
      responses:
        "200":
          description: Registered media asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"
  /media/{id}:
    get:
      tags: [media]
      summary: Get a media asset with a presigned download URL
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Media asset with presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAssetWithUrl"
        "404":
          description: Asset not found
    delete:
      tags: [media]
      summary: Delete a media asset
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "404":
          description: Asset not found
components:
  schemas:
...
    District:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        state:
          type: ["string", "null"]
        districtCode:
          type: ["string", "null"]
        adminId:
          type: ["string", "null"]
        schoolCount:
          type: integer
          description: Number of schools in this district.
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateDistrictBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        state:
          type: string
        districtCode:
          type: string
      required:
        - name
    School:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        state:
          type: ["string", "null"]
        schoolCode:
          type: ["string", "null"]
        districtId:
          type: ["string", "null"]
        districtName:
          type: ["string", "null"]
          description: Resolved district name (joined from districts table).
        teacherCount:
          type: integer
          description: Number of teachers linked to this school.
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    MediaAsset:
      type: object
      required: [id, name, s3Key, contentType, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        s3Key:
          type: string
        contentType:
          type: string
        size:
          type: integer
          nullable: true
        metadata:
          type: object
          additionalProperties: true
          nullable: true
        uploadedBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
    MediaAssetWithUrl:
      allOf:
        - $ref: "#/components/schemas/MediaAsset"
        - type: object
          required: [presignedUrl]
          properties:
            presignedUrl:
              type: string
              format: uri
    CreateMediaAssetBody:
      type: object
      required: [name, s3Key, contentType]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 1000
        s3Key:
          type: string
        contentType:
          type: string
        size:
          type: integer
        metadata:
          type: object
          additionalProperties: true
    MediaAssetList:
      type: object
      required: [assets, pagination]
      properties:
        assets:
          type: array
          items:
            $ref: "#/components/schemas/MediaAsset"
        pagination:
          type: object
          required: [page, limit, total, totalPages]
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer